+++ examples/src/main/scala/soda/example/FizzBuzz.scala
@@ -10,7 +10,7 @@ trait FizzBuzz {
   lazy val fizz_buzz =
     Recursion_ ()
       .range (100 )
-      .map (x => x + 1 )
+      .map ((x: Int ) => x + 1 )
       .map (get_fizz_buzz_term )
 
   def get_fizz_buzz_term (n: Int ) =